+++ snowav/utils/get_topo_stats.py
@@ -1,41 +1,29 @@
 
 import os
 import netCDF4 as nc
-from spatialnc import ipw
 
-def get_topo_stats(fp, filetype='netcdf'):
+def get_topo_stats(path):
     """
-    Get stats about topo from the topo file
-    Returns:
-        ts - dictionary of topo header data
+    Get stats about topo from the topo file.
+
+    Returns
+    ------
+    data - dictionary of topo header data
     """
 
-    fp = os.path.abspath(fp)
+    path = os.path.abspath(path)
+    data = nc.Dataset(path, 'r')
 
-    ts = {}
+    stats = {}
+    stats['y'] = data.variables['y'][:]
+    stats['x'] = data.variables['x'][:]
+    stats['nx'] = len(stats['x'])
+    stats['ny'] = len(stats['y'])
+    stats['du'] = stats['y'][1] - stats['y'][0]
+    stats['dv'] = stats['x'][1] - stats['x'][0]
+    stats['v'] = stats['x'][0]
+    stats['u'] = stats['y'][0]
 
-    if filetype == 'netcdf':
-        ds = nc.Dataset(fp, 'r')
-        # ts['units'] = ds.variables['y'].units
-        ts['y'] = ds.variables['y'][:]
-        ts['x'] = ds.variables['x'][:]
-        ts['nx'] = len(ts['x'])
-        ts['ny'] = len(ts['y'])
-        ts['du'] = ts['y'][1] - ts['y'][0]
-        ts['dv'] = ts['x'][1] - ts['x'][0]
-        ts['v'] = ts['x'][0]
-        ts['u'] = ts['y'][0]
-        ds.close()
+    data.close()
 
-    if filetype == 'ipw':
-        i = ipw.IPW(fp)
-        ts['nx'] = i.nsamps
-        ts['ny'] = i.nlines
-        ts['units'] = i.bands[0].units
-        ts['du'] = i.bands[0].dline
-        ts['dv'] = i.bands[0].dsamp
-        ts['v'] = float(i.bands[0].bsamp)
-        ts['u'] = float(i.bands[0].bline)
-        ts['csys'] = i.bands[0].coord_sys_ID
-
-    return ts
+    return stats